Include Linux flashlight-effect workaround
For some reason BorderColor-supports seems to be advertised although it
apparently is not (fully?) supported. At least on my system the
flashlight doesn't draw completely using the bordered texture.
To work around the issue where it exists this adds a ConVar to forcibly
disable the bordered version.

EZ2 doesn't seem to ship a _border material, causing a missing texture
to be projected, so default-enable the workaround there.